Split out getThemeForumUrl
